requirements.txt
python-dotenv==1.0.1
typer==0.12.5
httpx==0.27.2
pydantic==2.9.2
pytz==2024.1
python-dateutil==2.9.0.post0
dateparser==1.2.0

# Core math
numpy==1.26.4

# sklearn works with numpy 1.26
scikit-learn==1.4.2

# Remove faiss-cpu (it breaks on Windows + numpy 2.x)
# If you still want RAG, we use sklearn cosine similarity — that's enough

openai==1.53.0
anthropic==0.36.0
groq